% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvmmat_simu.R
\name{rsvmmat_simu}
\alias{rsvmmat_simu}
\title{Simulation for RSVMMAT test}
\usage{
rsvmmat_simu(
  n.sample = 1000,
  n.time = 5,
  par = list(),
  time_cov = TRUE,
  gene.count = 10,
  intercept = TRUE,
  causalpara = list(),
  power = FALSE,
  phe.model = "logistic",
  oversampling = "random"
)
}
\arguments{
\item{n.sample}{Numeric, sample size, number of individuals}

\item{n.time}{Numeric, number of measurements for each individual}

\item{par}{List, the parameters for the phenotype traits, including covaraites and individual specific time dependent random effects}

\item{time_cov}{Logical variable, indicating whether time effect is included in phenotypic traits}

\item{gene.count}{Numeric, number of genes}

\item{intercept}{Logical variable, indicating whether intercept is used in phenotypic traits}

\item{causalpara}{List, the parameters for causal genes}

\item{power}{Logical variable, indicating whether include disease genes in the generated genes}

\item{phe.model}{String, the phenotype model, two optional values: 'logistic', 'liability'}

\item{oversampling}{String, the ascertainment scheme, three optional value: 'random', 'baseline', 'sum'}
}
\value{
A list object is returned to be used as object for RSVMMAT test
}
\description{
This function use pre-defined parameters to make the simulation data for the RSVMMAT test (including type I and power test)
}
