% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvmmat_pvalue.R
\name{rsvmmat_test}
\alias{rsvmmat_test}
\title{Calculate prospective and retrospective P-values}
\usage{
rsvmmat_test(rsvmmat.est, G, impute.method = "fixed", GRM = NULL)
}
\arguments{
\item{rsvmmat.est}{The output of function "rsvmmat_est()"}

\item{G}{The genotype matrix, an m*q matrix where m is the number of subjects and q is the total number genetic variants.}

\item{impute.method}{choose the iputation method when there is missing genotype. Optional options are: 'random', 'fixed' or 'bestguess'.}

\item{GRM}{takes m-by-m genetic correlation matrix or kinship matrix.}
}
\value{
This function returns a vector of the pvalue of SVMMAT-B, SVMMAT-S,SVMMAT-A, RSVMMAT-B, RSVMMAT-S, and RSVMMAT-A.
}
\description{
This function tests a SNPs for a given SNP set for a given rsvmmat estimated null model.
}
